clean up route files, service and gateway route usage
